Add react router express middleware
