+++ server/middleware/dev-middleware.js
@@ -1,7 +1,9 @@
 import webpack from 'webpack';
 import webpackDevMiddleware from 'webpack-dev-middleware';
 import webpackHotMiddleware from 'webpack-hot-middleware';
+
 import base from '../../src/base/';
+import routingMiddleware from './routing-middleware';
 
 const config = require('../../webpack/index.babel.js');
